### 2018-03-01 #599 @whedon
[PRE REVIEW]: NiaPy: Python microframework for building nature-inspired algorithms
**Submitting author:** @GregaVrbancic (<a href="http://orcid.org/0000-0003-0723-3889">Grega Vrbančič</a>)
**Repository:** <a href="https://github.com/NiaOrg/NiaPy" target ="_blank">https://github.com/NiaOrg/NiaPy</a>
**Version:** 1.0.0
