### 2018-03-01 26d75a5
fix paper styling
