+++ paper/paper.bib
@@ -16,7 +16,8 @@
   volume  = {4},
   number  = {2},
   year    = {2013},
-  pages   = {150--194}
+  pages   = {150--194},
+  doi     = {10.1504/IJMMNO.2013.055204}
 }
 
 @article{Sorensen2015,
@@ -27,7 +28,8 @@
   number    = {1},
   pages     = {3--18},
   year      = {2015},
-  publisher = {Wiley Online Library}
+  publisher = {Wiley Online Library},
+  doi       = {10.1111/itor.12001}
 }
 
 @book{Yang2014,
