### 2018-03-08 #599 @ljvmiranda921
Hi, I was also contacted by one of the authors (Dr. Iztok Fister Jr.) through e-mail  to review this submission. I'm interested to review but:
- I'll only be free by March 19; and
- I'm not part of the official reviewers of JOSS

In addition, I can also recommend @\stsievert , my reviewer during [my submission in PySwarms](https://github.com/openjournals/joss-reviews/issues/433) if that will be a better option.
