### 2018-03-11 #599 @arfon
@ljvmiranda921 - excellent. Thanks!

> I'm not part of the official reviewers of JOSS

Absolutely, there's no 'requirement' to be an official reviewer for future reviews.
