### 2018-03-19 NiaOrg/NiaPy#119 @ljvmiranda921
[JOSS] Clarify set-up requirements in README and requirements.txt
1. It would be better if the required Python modules are clarified in your README.  As of now, I can only see `python 3.6+ (2.7.14)` and `pip`. But your [`setup.py`](https://github.com/NiaOrg/NiaPy/blob/master/setup.py) seems to be installing:

- numpy <= 1.14
- click <= 6.0
- scipy <= 1.0.0
- xlsxwriter <= 1.0.2

2. In addition, having `setup.py` install from a `requirements.txt` may be more preferred that defining them individually inside the script. [Check this out](https://pip.readthedocs.io/en/1.1/requirements.html). Although I believe the `Pipfile` solves this, so this is optional.

3. Lastly, it is also beneficial to separate "user requirements" and "dev requirements." The latter may include testing libraries you've used (pytest), and sphinx for building the documentation. You can just link [this](http://niapy.readthedocs.io/en/latest/installation.html) to your requirements header
---
This is an issue related to openjournals/joss-reviews#613
