### 2018-03-19 NiaOrg/NiaPy#120 @ljvmiranda921
[JOSS] Better API Documentation
1. Stylistically, it may be better to pattern the docstring with the standard template found [here](https://pythonhosted.org/an_example_pypi_project/sphinx.html#function-definitions). This is so that it will render properly in your [API Documentation](http://niapy.readthedocs.io/en/latest/api/algorithms.html#module-NiaPy.algorithms.basic).

Another option is to use the [`numpydoc` extension](https://github.com/numpy/numpydoc), which gives a prettier format inside your docstrings.

2. It's also useful to add docstrings on some of your class methods. For example, [this method](http://niapy.readthedocs.io/en/stable/_modules/NiaPy/algorithms/basic/ga.html#GeneticAlgorithm.checkForBest) doesn't have any docstring/documentation.
---
This is an issue related to openjournals/joss-reviews#613
