### 2018-03-19 NiaOrg/NiaPy#121 @ljvmiranda921
[JOSS] Comment on existing libraries/frameworks
There are already implementations of nature-inspired algorithms/frameworks in Python. To name a few, we have [DEAP](https://github.com/DEAP/deap), [Evolopy](https://github.com/7ossam81/EvoloPy), and [Inspyred](https://github.com/aarongarrett/inspyred). It may be worth acknowledging some of these frameworks.

Furthermore, it may also be necessary to expound upon how NiaPy differs from other libraries, does it solve a different problem or use a different software architecture? You can leverage on the algorithms you've implemented that doesn't exist yet in other works, and comment on Niapy in the context of "white-box" or "black-box" approach on software implementation ([ref1](http://vision.gel.ulaval.ca/~cgagne/pubs/deap-gecco-2012.pdf), [ref2](https://github.com/jakobbossek/ecr2#black-box-approach)).

---
This is an issue related to openjournals/joss-reviews#613
