### 2018-03-19 NiaOrg/NiaPy#122 @ljvmiranda921
[JOSS] Clarify target audience
Please write a sentence or two on who are the intended audience for this library. I know it's pretty obvious for us in this field, but it will be better to make it explicit in the paper.

---
This is an issue related to openjournals/joss-reviews#613
