### 2018-03-19 #613 @ljvmiranda921
Hello @GregaVrbancic , I have opened-up some issues from my review, please check them out.
