### 2018-03-19 #613 @GregaVrbancic
Hello @ljvmiranda921 , thank you for your time and effort to review our software. We'll take a look at opened issues and will try to address them as soon as possible.
