### 2018-03-19 NiaOrg/NiaPy#123 @ljvmiranda921
[JOSS] (Optional) Follow PEP-8 style guide in naming methods
This is prescriptive and optional, but I personally find this very important. Please follow [PEP8](https://www.python.org/dev/peps/pep-0008/#package-and-module-names) as a prescriptive guide.

- Methods usually have lower-case characters separated by underscores. `TournamentSelection` should be `tournament_selection`. Check [this](https://www.python.org/dev/peps/pep-0008/#function-and-variable-names).
- The package name should be in lower-case. `NiaPy` should be `niapy`. Check [this](https://www.python.org/dev/peps/pep-0008/#package-and-module-names).

---
This is an issue related to openjournals/joss-reviews#613
