### 2018-03-19 35046d3
add dependencies list to README with link to development dependencies list.
