+++ README.rst
@@ -96,6 +96,16 @@ Requirements
 -  Python 3.6+ (backward compatibility with 2.7.14)
 -  Pip
 
+Dependencies
+~~~~~~~~~~~~
+
+-  click == *
+-  numpy == 1.14.0
+-  scipy == 1.0.0
+-  xlsxwriter == 1.0.2
+
+List of development dependencies and requirements can be found in the :doc:`/installation`.
+
 Installation
 ------------
 
