### 2018-03-19 NiaOrg/NiaPy#119 @GregaVrbancic
Firstly, thank you very much for such a detailed description of an issue!

We have have clarified required Python modules for NiaPy in README file. Also we have added link to detailed list of dependency modules and development dependency modules as recommended.

Please check out and confirm the changes we made in referenced PR #124 before we merge them into master branch.

Thank you!
