### 2018-03-19 NiaOrg/NiaPy#119 @ljvmiranda921
Hi @GregaVrbancic ,

Looks good to me!
