### 2018-03-19 NiaOrg/NiaPy#119 @GregaVrbancic
Thank you!
