### 2018-03-19 NiaOrg/NiaPy#121 @firefly-cpp
Dear @ljvmiranda921,

Thank you very much for your detailed comments! The following projects will be mentioned in our paper:

DEAP

@article{deap,
    author    = " F\'elix-Antoine Fortin and Fran\c{c}ois-Michel {De Rainville} and Marc-Andr\'e Gardner and Marc Parizeau and Christian Gagn\'e ",
    title     = { {DEAP}: Evolutionary Algorithms Made Easy },
    pages    = { 2171--2175 },
    volume    = { 13 },
    year      = { 2012 },
    journal   = { Journal of Machine Learning Research }
}

EvoloPy

@inproceedings{evolopy,
  title={EvoloPy: An Open-source Nature-inspired Optimization Framework in Python.},
  author={Faris, Hossam and Aljarah, Ibrahim and Mirjalili, Seyedali and Castillo, Pedro A and Merelo, Juan J},
  booktitle={Proceedings of the 8th International Joint Conference on Computational Intelligence},
  pages={171--177},
  year={2016}
}

PySwarms

@article{pyswarms,
    author  = {Lester James V. Miranda},
    title   = "{P}y{S}warms, a research-toolkit for {P}article {S}warm {O}ptimization in {P}ython",
    journal = {Journal of Open Source Software},
    year    = {2018},
    volume  = {3},
    issue   = {21},
    doi     = {10.21105/joss.00433},
    url     = {https://doi.org/10.21105/joss.00433}
}

inspyred

@article{inspyred,
  title={inspyred: Bio-inspired Algorithms in Python},
  author={Garrett, Aaron},
  journal={URL: https://pypi. python. org/pypi/inspyred},
  year={2014}
}

Pygmo

@inproceedings{pygmo,
  title={Pygmo and pykep: Open source tools for massively parallel optimization in astrodynamics (the case of interplanetary trajectory optimization)},
  author={Izzo, Dario},
  booktitle={Proceedings of the Fifth International Conference on Astrodynamics Tools and Techniques, ICATT},
  year={2012}
}

Do you have any additional suggestions?
