### 2018-03-20 NiaOrg/NiaPy#121 @ljvmiranda921
Hi, this looks good already. Also, please add a sentence or two in the paper describing how Niapy relates to these frameworks. Doesn't need to be a full-blown RRL :+1:
