+++ paper/paper.bib
@@ -45,3 +45,46 @@
   year      = {2016},
   publisher = {CRC Press}
 }
+
+@article{deap,
+  title={DEAP: Evolutionary algorithms made easy},
+  author={Fortin, F{\'e}lix-Antoine and Rainville, Fran{\c{c}}ois-Michel De and Gardner, Marc-Andr{\'e} and Parizeau, Marc and Gagn{\'e}, Christian},
+  journal={Journal of Machine Learning Research},
+  volume={13},
+  number={Jul},
+  pages={2171--2175},
+  year={2012}
+}
+
+@inproceedings{evolopy,
+  title={EvoloPy: An Open-source Nature-inspired Optimization Framework in Python.},
+  author={Faris, Hossam and Aljarah, Ibrahim and Mirjalili, Seyedali and Castillo, Pedro A and Merelo, Juan J},
+  booktitle={Proceedings of the 8th International Joint Conference on Computational Intelligence},
+  pages={171--177},
+  year={2016}
+}
+
+@article{pyswarms,
+    author  = {Lester James V. Miranda},
+    title   = {PySwarms: a research toolkit for Particle Swarm Optimization in Python},
+    journal = {Journal of Open Source Software},
+    year    = {2018},
+    volume  = {3},
+    issue   = {21},
+    doi     = {10.21105/joss.00433},
+    url     = {https://doi.org/10.21105/joss.00433}
+}
+
+@article{inspyred,
+  title={inspyred: Bio-inspired Algorithms in Python},
+  author={Garrett, Aaron},
+  journal={URL: https://pypi. python. org/pypi/inspyred},
+  year={2014}
+}
+
+@inproceedings{pygmo,
+  title={Pygmo and pykep: Open source tools for massively parallel optimization in astrodynamics (the case of interplanetary trajectory optimization)},
+  author={Izzo, Dario},
+  booktitle={Proceedings of the Fifth International Conference on Astrodynamics Tools and Techniques, ICATT},
+  year={2012}
+}
