### 2018-03-20 NiaOrg/NiaPy#121 @firefly-cpp
Hi @ljvmiranda921 ,

The following text is going to be added in paper.

"There are also some existing similar Python implementations/frameworks of nature-inspired algorithms, as for example DEAP [@deap], EvoloPy [@evolopy], PySwarms [@pyswarms], inspyred [@inspyred] and Pygmo [@pygmo]. However, NiaPy tends to be very minimalistic and easy for use, especially for practitioners and students. "
