### 2018-03-20 2e65e61
Documentation added
Documentation of methods added
