### 2018-03-20 NiaOrg/NiaPy#120 @lucijabrezocnik
Hi @ljvmiranda921 ,

We have added missing documentation to the classes.

Regarding the style of docstring pattern, we revised your recommendation and concluded to update our docstring following one of the proposed options in one of the future releases.

Please check out and confirm the changes we made in referenced PR #125 before we merge them into the master branch.

Thank you!
