### 2018-03-20 NiaOrg/NiaPy#121 @firefly-cpp
This is also related to issue: #122
