### 2018-03-20 NiaOrg/NiaPy#121 @ljvmiranda921
This looks good to me, I appreciate citing PySwarms but the library has more of a niche scope (only swarm optimization and its variants) compared to others (and your work) that offers optimizers outside of swarm opt. You can drop it if you deem it unnecessary.

#121 and #122 can be closed once the branches are merged. :+1:
