### 2018-03-20 NiaOrg/NiaPy#120 @ljvmiranda921
> Regarding the style of docstring pattern, we revised your recommendation and concluded to update our docstring following one of the proposed options in one of the future releases.

No problem.

> We have added missing documentation to the classes.
> Please check out and confirm the changes we made in referenced PR #125 before we merge them into the master branch.

The descriptions can be more verbose in the future but this looks good to me already. I believe you can already merge to master.
