### 2018-03-20 #613 @ljvmiranda921
Hello @GregaVrbancic ,

I appreciate the swift response on the issues. Overall the library looks good once the remaining branches are merged. Once we close the issues, we can complete the checklist above and call-in the Editor.
