### 2018-03-20 NiaOrg/NiaPy#120 @lucijabrezocnik
Thank you for your fast response @ljvmiranda921 !

We will add more verbose descriptions in the next release 👍
