### 2018-03-20 #613 @GregaVrbancic
@ljvmiranda921 thank you again for such a quick responses on opened issues. We are working on closing those remaining opened issues. If I understood correctly, after closing those issues and completing the checklist above, we have to make a new release of software and archive it on zenodo?
