+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{Fister2013,
-  author  = {Fister Jr., Iztok and Yang, Xin-She and Fister, Iztok and Brest, Janez and Fister, Du{\v{s}}an},
+  author  = {Iztok Fister Jr., Xin-She Yang, Iztok Fister, Janez Brest, and  Du{\v{s}}an Fister},
   title   = {{A Brief Review of Nature-Inspired Algorithms for Optimization}},
   journal = {Elektrotehni\v{s}ki vestnik},
   number  = {3},
@@ -9,7 +9,7 @@
 }
 
 @article{Jamil2013,
-  author  = {Jamil, Momin and Yang, Xin-She},
+  author  = {Momin Jamil and Xin-She Yang},
   title   = {{A Literature Survey of Benchmark Functions For Global Optimization
              Problems}},
   journal = {International Journal of Mathematical Modelling and Numerical Optimisation},
@@ -21,7 +21,7 @@
 }
 
 @article{Sorensen2015,
-  author    = {S{\"o}rensen, Kenneth},
+  author    = {Kenneth S{\"o}rensen},
   title     = {{Metaheuristics—the metaphor exposed}},
   journal   = {International Transactions in Operational Research},
   volume    = {22},
@@ -33,58 +33,57 @@
 }
 
 @book{Yang2014,
-  author    = {Yang, Xin-She},
+  author    = {Xin-She Yang},
   title     = {{Nature-inspired optimization algorithms}},
   year      = {2014},
   publisher = {Elsevier}
 }
 
 @book{Hassanien2016,
-  author    = {Hassanien, Aboul Ella and Emary, Eid},
+  author    = {Aboul Ella Hassanien and Eid Emary},
   title     = {{Swarm intelligence: principles, advances, and applications}},
   year      = {2016},
   publisher = {CRC Press}
 }
 
-@article{deap,
-  title={DEAP: Evolutionary algorithms made easy},
-  author={Fortin, F{\'e}lix-Antoine and Rainville, Fran{\c{c}}ois-Michel De and Gardner, Marc-Andr{\'e} and Parizeau, Marc and Gagn{\'e}, Christian},
-  journal={Journal of Machine Learning Research},
-  volume={13},
-  number={Jul},
-  pages={2171--2175},
-  year={2012}
+@article{Deap2012,
+  author    = {F{\'e}lix-Antoine Fortin, Fran{\c{c}}ois-Michel De Rainville, Marc-Andr{\'e} Gardner, Marc Parizeau, and Christian Gagn{\'e}},
+  title     = {{DEAP: Evolutionary algorithms made easy}},
+  journal   = {Journal of Machine Learning Research},
+  volume    = {13},
+  number    = {Jul},
+  pages     = {2171--2175},
+  year      = {2012}
 }
 
-@inproceedings{evolopy,
-  title={EvoloPy: An Open-source Nature-inspired Optimization Framework in Python.},
-  author={Faris, Hossam and Aljarah, Ibrahim and Mirjalili, Seyedali and Castillo, Pedro A and Merelo, Juan J},
-  booktitle={Proceedings of the 8th International Joint Conference on Computational Intelligence},
-  pages={171--177},
-  year={2016}
+@inproceedings{Evolopy2016,
+  author    = {Hossam Faris, Ibrahim Aljarah, Seyedali Mirjalili, Pedro A Castillo, and Juan J Merelo},
+  title     = {{EvoloPy: An Open-source Nature-inspired Optimization Framework in Python.}},
+  booktitle = {Proceedings of the 8th International Joint Conference on Computational Intelligence},
+  pages     = {171--177},
+  year      = {2016}
 }
 
-@article{pyswarms,
-    author  = {Lester James V. Miranda},
[…]
