### 2018-03-20 NiaOrg/NiaPy#123 @GregaVrbancic
Thank you for recommendations and provided links to resources. We will most definitely go through PEP8 - code style guide and refactor our code in one of the next releases, to be in line with best practices and style conventions.
