### 2018-03-20 #613 @ljvmiranda921
> If I understood correctly, after closing those issues and completing the checklist above, we have to make a new release of software and archive it on zenodo?

Yes I believe that's correct. But I'd prefer we wait for the editor's go signal first before making a new release so we won't miss anything and avoid making multiple releases.
