### 2018-03-20 NiaOrg/NiaPy#121 @firefly-cpp
Dear @ljvmiranda921,

I really believe that PySwarms framework has a place in related work section. It is a good example of
a well-documented and extensible framework, where other SI algorithms can also be easily developed within the framework.

Thanks again for your quick responses.
