### 2018-03-21 #613 @ljvmiranda921
Hi @arfon ,

I am done reviewing NiaPy. The issues I opened (referenced here in this issue) have now been closed. All boxes have been checked.

## Review Summary

NiaPy is a Python library for nature-inspired optimization. It offers a simple and easy-to-use solution to minimize algorithm implementation from scratch. Overall, the software is a good candidate for publication in JOSS.

### Functionality
The software functions as claimed in the [basic examples](http://niapy.readthedocs.io/en/latest/getting_started.html#basic-example) and guides. There are no performance claims so I checked that one off. I've also played around and tested their `ParticleSwarmOptimization` classes and it works as expected.

### Documentation
The installation instructions are clear and works on my Linux (Ubuntu 16.04) and Windows 10 machines as stated. The dependencies, together with the dev requirements, are also listed in both the README and Documentation. API documentation for both methods and classes are also included. Overall, the software works as advertised in the documentation.

There is a clear statement of need in the README. The [contributing guidelines](http://niapy.readthedocs.io/en/latest/guides/index.html) are complete (development, testing, git collab) and can be found in the documentation. A [code of conduct](https://github.com/NiaOrg/NiaPy/blob/master/CODE_OF_CONDUCT.md) is also included.

### Paper
The paper is clear and gives a good overview of the problem, a statement of need, and a short survey of existing works. The library architecture is also included, which I consider as a plus. The references are also complete.

---
#### Recommendation for future releases

Although I considered it optional, I hope that class methods and package names will follow the [PEP8 style](https://www.python.org/dev/peps/pep-0008/). PEP8 is prescriptive, but it is a good standard for Python projects, especially for something useful as NiaPy. However, changing everything into PEP8 will most likely introduce breaking changes, so it's up to the discretion of the authors to follow this suggestion.

---
