+++ README.rst
@@ -104,7 +104,7 @@ Dependencies
 -  scipy == 1.0.0
 -  xlsxwriter == 1.0.2
 
-List of development dependencies and requirements can be found in the :doc:`/installation`.
+List of development dependencies and requirements can be found in the `installation section of NiaPy documentation <http://niapy.readthedocs.io/en/stable/installation.html>`_.
 
 Installation
 ------------
@@ -165,6 +165,13 @@ fits your purposes or that it is bug-free. Use it at your own risk!
 
 Revision History
 ================
+
+1.0.1 (Mar 21, 2018)
+--------------------
+This release reflects the changes from Journal of Open Source Software (JOSS) review:
+- Better API Documentation
+- Clarification of set-up requirements in README
+- Improved paper
 
 1.0.0 (Feb 28, 2018)
 --------------------
