+++ docs/source/changelog.rst
@@ -3,6 +3,13 @@ Changelog
 
 We are using `semantic versioning <https://semver.org/>`_.
 
+1.0.1 (Mar 21, 2018)
+--------------------
+This release reflects the changes from Journal of Open Source Software (JOSS) review:
+- Better API Documentation
+- Clarification of set-up requirements in README
+- Improved paper
+
 1.0.0 (Feb 28, 2018)
 --------------------
 - stable release 1.0.0
@@ -12,7 +19,7 @@ We are using `semantic versioning <https://semver.org/>`_.
 - fix PyPI build
 
 1.0.0rc1 (Feb 28, 2018)
--------
+-----------------------
 - version 1.0.0 release candidate 1
 - added 10 algorithms
 - added 26 benchmark functions
