### 2018-03-21 #613 @lucijabrezocnik
Hi @arfon !

We have prepared a new release and archived the current version of NiaPy on Zenodo ([https://zenodo.org/record/1205048](https://zenodo.org/record/1205048)). Here is the DOI: 10.5281/zenodo.1205048.
